+++ app/src/main/java/ch/heigvd/pro/b04/android/Question/Answer/AnswerAdapter.java
@@ -1,4 +1,4 @@
-package ch.heigvd.pro.b04.android.Poll.Answer;
+package ch.heigvd.pro.b04.android.Question.Answer;
 
 import android.graphics.Color;
 import android.view.LayoutInflater;
@@ -13,7 +13,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.LinkedList;
 import java.util.List;
 
-import ch.heigvd.pro.b04.android.Poll.Question.QuestionViewModel;
+import ch.heigvd.pro.b04.android.Question.QuestionViewModel;
 import ch.heigvd.pro.b04.android.R;
 
 public class AnswerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>  {
