+++ app/src/main/java/ch/heigvd/pro/b04/android/Question/QuestionViewModel.java
@@ -1,4 +1,4 @@
-package ch.heigvd.pro.b04.android.Poll.Question;
+package ch.heigvd.pro.b04.android.Question;
 
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
@@ -7,7 +7,7 @@ import java.util.LinkedList;
 import java.util.List;
 
 import ch.heigvd.pro.b04.android.Datamodel.Question;
-import ch.heigvd.pro.b04.android.Poll.Answer.Answer;
+import ch.heigvd.pro.b04.android.Question.Answer.Answer;
 
 public class QuestionViewModel extends ViewModel {
     private MutableLiveData<Question> question = new MutableLiveData<>();
