+++ app/src/main/res/layout/activity_question.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ch.heigvd.pro.b04.android.Poll.Question.QuestionActivity">
+    tools:context=".Question.QuestionActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@id/question_answers_view"
